Some changes to function structure
The arrays of players and slots were declared inside the fromUser function and this caused problems. On top of declaring these arrays, I decided to split this function into two. This makes the code a little cleaner and also makes it easier to define the arrays in main. The array of slots is now of type enum. Function populateBoard doesn't want to work. Reading from files doesn't work yet.